+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - PATCH Sound no longer plays when the MainForm opens (caused by setting the useanalog checked event link and then setting the value to loaded data which triggered it).
 - MINOR Skips checking for a new version if there is no internet connection.
 
-## [0.0.0.0] - 2019-06-XX
+## IN PROGRESS
 ### Added
 - MINOR "ACTIONS PERFORMED EVERYTIME AT LAUNCH" text added to the LAUNCH OPTIONS menu to explain what each option does to avoid appending "on play" to the end of every option.
 - MINOR Button that opens the download page for Livesplit.